+++ src/client/components/DemographicsTooltip.tsx
@@ -1,6 +1,6 @@
 /** @jsx jsx */
 import mapValues from "lodash/mapValues";
-import { Box, jsx, Styled, ThemeUIStyleObject } from "theme-ui";
+import { Box, jsx, Styled, ThemeUIStyleObject, Divider } from "theme-ui";
 
 import { demographicsColors } from "../constants/colors";
 
@@ -9,6 +9,7 @@ const style: ThemeUIStyleObject = {
     textAlign: "left",
     py: 0,
     pr: 2,
+    whiteSpace: "nowrap",
     textTransform: "capitalize"
   },
   number: {
@@ -19,6 +20,10 @@ const style: ThemeUIStyleObject = {
     fontWeight: "light"
   }
 };
+
+function parseRowLabel(label: string) {
+  return label.split(/(?=[A-Z])/).join(" ");
+}
 
 const Row = ({
   label,
@@ -35,7 +40,7 @@ const Row = ({
       border: "none"
     }}
   >
-    <Styled.td sx={style.label}>{label}</Styled.td>
+    <Styled.td sx={style.label}>{parseRowLabel(label)}</Styled.td>
     <Styled.td sx={{ minWidth: "50px", py: 0 }}>
       <Box
         style={{
@@ -56,24 +61,42 @@ const Row = ({
 );
 
 const DemographicsTooltip = ({
-  demographics
+  demographics,
+  isMinorityMajority
 }: {
   readonly demographics: { readonly [id: string]: number };
+  readonly isMinorityMajority?: boolean;
 }) => {
   const percentages = mapValues(
     demographics,
     (population: number) =>
       (demographics.population ? population / demographics.population : 0) * 100
   );
-  const races = ["white", "black", "asian", "hispanic", "other"] as const;
-  const rows = races.map((id: typeof races[number]) => (
-    <Row key={id} label={id} percent={percentages[id]} color={demographicsColors[id]} />
-  ));
+  const races = [
+    "white",
+    "black",
+    "asian",
+    "hispanic",
+    "nativeAmerican",
+    "pacificIslander",
+    "other"
+  ] as const;
+  const rows = races
+    .filter(race => percentages[race] !== undefined)
+    .map((id: typeof races[number]) => (
+      <Row key={id} label={id} percent={percentages[id]} color={demographicsColors[id]} />
+    ));
   return (
     <Box sx={{ width: "100%", minHeight: "100%" }}>
       <Styled.table sx={{ margin: "0", width: "100%" }}>
         <tbody>{rows}</tbody>
       </Styled.table>
+      {isMinorityMajority && (
+        <Box>
+          <Divider sx={{ my: 1, borderColor: "gray.6" }} />
+          <Box>* Minority majority district</Box>
+        </Box>
+      )}
     </Box>
   );
 };