+++ src/client/components/MapHeader.tsx
@@ -1,12 +1,8 @@
 /** @jsx jsx */
+import React from "react";
 import { Flex, Box, Label, Button, jsx, Select, ThemeUIStyleObject } from "theme-ui";
 import { GeoLevelInfo, GeoLevelHierarchy, GeoUnits, IStaticMetadata } from "../../shared/entities";
-import {
-  areAnyGeoUnitsSelected,
-  geoLevelLabel,
-  isBaseGeoLevelAlwaysVisible,
-  capitalizeFirstLetter
-} from "../functions";
+import { geoLevelLabel, capitalizeFirstLetter, canSwitchGeoLevels } from "../functions";
 import MapSelectionOptionsFlyout from "./MapSelectionOptionsFlyout";
 
 import Icon from "./Icon";
@@ -15,7 +11,6 @@ import {
   setGeoLevelIndex,
   setSelectionTool,
   SelectionTool,
-  showAdvancedEditingModal,
   setMapLabel
 } from "../actions/districtDrawing";
 import store from "../store";
@@ -87,7 +82,6 @@ const GeoLevelButton = ({
   geoLevelIndex,
   geoLevelHierarchy,
   selectedGeounits,
-  advancedEditingEnabled,
   isReadOnly
 }: {
   readonly index: number;
@@ -95,24 +89,13 @@ const GeoLevelButton = ({
   readonly geoLevelIndex: number;
   readonly geoLevelHierarchy: GeoLevelHierarchy;
   readonly selectedGeounits: GeoUnits;
-  readonly advancedEditingEnabled?: boolean;
   readonly isReadOnly: boolean;
 }) => {
   const label = geoLevelLabel(value.id);
-  const areGeoUnitsSelected = areAnyGeoUnitsSelected(selectedGeounits);
-  const isBaseLevelAlwaysVisible = isBaseGeoLevelAlwaysVisible(geoLevelHierarchy);
-  const isBaseGeoLevelSelected = geoLevelIndex === geoLevelHierarchy.length - 1;
-  const isCurrentLevelBaseGeoLevel = index === geoLevelHierarchy.length - 1;
-  const areChangesPending =
-    !isBaseLevelAlwaysVisible &&
-    areGeoUnitsSelected &&
-    // block level selected, so disable all higher geolevels
-    ((isBaseGeoLevelSelected && !isCurrentLevelBaseGeoLevel) ||
-      // non-block level selected, so disable block level
-      (!isBaseGeoLevelSelected && isCurrentLevelBaseGeoLevel));
-  // Always show the currently selected geolevel, even if it would otherwise be hidden
+  const canSwitch = canSwitchGeoLevels(geoLevelIndex, index, geoLevelHierarchy, selectedGeounits);
+  // Always show the currently selected geolevel, even if it would otherwise be disabled
   const isCurrentLevelSelected = index === geoLevelIndex;
-  const isButtonDisabled = !isCurrentLevelSelected && areChangesPending;
+  const isButtonDisabled = !isCurrentLevelSelected && !canSwitch;
 
   return (
     <Box sx={{ ...style.buttonGroup, ...{ display: "inline-block", position: "relative" } }}>
@@ -127,16 +110,7 @@ const GeoLevelButton = ({
             key={index}
             sx={{ ...style.selectionButton, ...{ mr: "1px" } }}
             className={buttonClassName(geoLevelIndex === index)}
-            onClick={() =>
-              store.dispatch(
-                !isCurrentLevelBaseGeoLevel ||
-                  isReadOnly ||
-                  advancedEditingEnabled ||
-                  isBaseLevelAlwaysVisible
-                  ? setGeoLevelIndex(index)
-                  : showAdvancedEditingModal(true)
-              )
-            }
+            onClick={() => store.dispatch(setGeoLevelIndex({ index, isReadOnly }))}
             disabled={isButtonDisabled}
           >
             {label}
@@ -153,7 +127,6 @@ const MapHeader = ({
   selectionTool,
   geoLevelIndex,
   selectedGeounits,
-  advancedEditingEnabled,
   isReadOnly,
   limitSelectionToCounty
 }: {
@@ -188,7 +161,6 @@ const MapHeader = ({
             geoLevelIndex={geoLevelIndex}
             geoLevelHierarchy={geoLevelHierarchy}
             selectedGeounits={selectedGeounits}
-            advancedEditingEnabled={advancedEditingEnabled}
             isReadOnly={isReadOnly}
           />
         ))
@@ -197,42 +169,45 @@ const MapHeader = ({
     <Flex sx={style.header}>
       <Flex>
         {!isReadOnly && (
-          <Flex sx={{ ...style.buttonGroup, mr: 2 }}>
-            <Tooltip content="Point-and-click selection">
-              <Button
-                sx={{ ...style.selectionButton }}
-                className={buttonClassName(selectionTool === SelectionTool.Default)}
-                onClick={() => store.dispatch(setSelectionTool(SelectionTool.Default))}
-              >
-                <Icon name="hand-pointer" />
-              </Button>
-            </Tooltip>
-            <Tooltip content="Rectangle selection">
-              <Button
-                sx={{ ...style.selectionButton }}
-                className={buttonClassName(selectionTool === SelectionTool.Rectangle)}
-                onClick={() => store.dispatch(setSelectionTool(SelectionTool.Rectangle))}
-              >
-                <Icon name="draw-square" />
-              </Button>
-            </Tooltip>
-            <Tooltip content="Paint brush selection">
-              <Button
-                sx={{ ...style.selectionButton }}
-                className={buttonClassName(selectionTool === SelectionTool.PaintBrush)}
-                onClick={() => store.dispatch(setSelectionTool(SelectionTool.PaintBrush))}
-              >
-                <Icon name="paint-brush" />
-              </Button>
-            </Tooltip>
-          </Flex>
+          <React.Fragment>
+            <Flex sx={{ ...style.buttonGroup, mr: 2 }}>
+              <Tooltip content="Point-and-click selection">
+                <Button
+                  sx={{ ...style.selectionButton }}
+                  className={buttonClassName(selectionTool === SelectionTool.Default)}
+                  onClick={() => store.dispatch(setSelectionTool(SelectionTool.Default))}
+                >
+                  <Icon name="hand-pointer" />
+                </Button>
+              </Tooltip>
+              <Tooltip content="Rectangle selection">
+                <Button
+                  sx={{ ...style.selectionButton }}
+                  className={buttonClassName(selectionTool === SelectionTool.Rectangle)}
+                  onClick={() => store.dispatch(setSelectionTool(SelectionTool.Rectangle))}
+                >
+                  <Icon name="draw-square" />
+                </Button>
+              </Tooltip>
+              <Tooltip content="Paint brush selection">
+                <Button
+                  sx={{ ...style.selectionButton }}
+                  className={buttonClassName(selectionTool === SelectionTool.PaintBrush)}
+                  onClick={() => store.dispatch(setSelectionTool(SelectionTool.PaintBrush))}
+                >
+                  <Icon name="paint-brush" />
+                </Button>
+              </Tooltip>
+            </Flex>
+
+            <Box sx={{ position: "relative", mr: 3, pt: "6px" }}>
+              <MapSelectionOptionsFlyout
+                limitSelectionToCounty={limitSelectionToCounty}
+                topGeoLevelName={topGeoLevelName}
+              />
+            </Box>
+          </React.Fragment>
         )}
-        <Box sx={{ position: "relative", mr: 3, pt: "6px" }}>
-          <MapSelectionOptionsFlyout
-            limitSelectionToCounty={limitSelectionToCounty}
-            topGeoLevelName={topGeoLevelName}
-          />
-        </Box>
         <Flex className="geolevel-button-group">{geoLevelOptions}</Flex>
       </Flex>
       <Box sx={{ lineHeight: "1" }}>