+++ src/client/components/evaluate/ProjectEvaluateSidebar.tsx
@@ -23,7 +23,7 @@ const style: ThemeUIStyleObject = {
     flexShrink: 0,
     height: "100%",
     minWidth: "430px",
-    maxWidth: "447px",
+    maxWidth: "507px",
     position: "relative",
     zIndex: 200
   }
@@ -48,10 +48,11 @@ const ProjectEvaluateSidebar = ({
   const popThreshold = 0.05;
   useEffect(() => {
     if (geojson && !avgCompactness) {
-      const totalCompactness = geojson.features.reduce(function(accumulator, feature) {
+      const features = geojson.features.slice(1).filter(f => f.properties.compactness !== 0);
+      const totalCompactness = features.reduce(function(accumulator, feature) {
         return accumulator + feature.properties.compactness;
       }, 0);
-      setAvgCompactness(totalCompactness / (geojson.features.length - 1));
+      setAvgCompactness(features.length !== 0 ? totalCompactness / features.length : undefined);
     }
   }, [geojson, avgCompactness]);
 
@@ -80,12 +81,24 @@ const ProjectEvaluateSidebar = ({
     {
       key: "equalPopulation",
       name: "Equal Population",
-      status: false,
+      status:
+        (avgPopulation &&
+          geojson &&
+          geojson?.features.filter(f => {
+            return (
+              f.id !== 0 &&
+              f.properties.percentDeviation &&
+              Math.abs(f.properties.percentDeviation) <= popThreshold
+            );
+          }).length ===
+            geojson?.features.length - 1) ||
+        false,
       description: "have equal population",
-      shortText: "Lorem ipsum",
-      longText: `This map has a goal of getting within ${Math.floor(
+      shortText:
+        "The U.S. constitution requires that each district have about the same population for a map to be considered valid.",
+      longText: `Districts are required to be "Equal Population" for a map to be considered valid. Districts are "Equal Population" when their population falls within the target threshold. The target population is the total population divided by the number of districts and the threshold is a set percentage deviation (${Math.floor(
         popThreshold * 100
-      )}% of ${avgPopulation?.toLocaleString()}`,
+      )}%) above or below that target.`,
       avgPopulation: avgPopulation || undefined,
       popThreshold: popThreshold,
       type: "fraction",
@@ -95,9 +108,7 @@ const ProjectEvaluateSidebar = ({
           ? geojson?.features.filter(f => {
               return (
                 f.id !== 0 &&
-                // @ts-ignore
                 f.properties.percentDeviation &&
-                // @ts-ignore
                 Math.abs(f.properties.percentDeviation) <= popThreshold
               );
             }).length
@@ -110,9 +121,9 @@ const ProjectEvaluateSidebar = ({
         geojson?.features.filter(f => f.properties.contiguity === "non-contiguous").length === 0,
       description: "are contiguous",
       longText:
-        "A district must be in a single, unbroken shape. Two areas touching at the corners are typically not considered contiguous. An exception would be the inclusion of islands in a coastal district.",
+        "Each district should be contiguous, meaning it must be a single, unbroken shape. Some exceptions are allowed, such as the inclusion of islands in a coastal district. Two areas connected only by a single point (touching just their corners) are not considered contiguous.",
       shortText:
-        "All parts of a district must be in physical contact with some other part of the district",
+        "All parts of a district must be in physical contact with some other part of the district, to the extent possible.",
       type: "fraction",
       total: geojson?.features.filter(f => f.id !== 0).length || 0,
       value:
@@ -121,40 +132,44 @@ const ProjectEvaluateSidebar = ({
     }
   ];
   const optionalMetrics: readonly EvaluateMetric[] = [
-    {
-      key: "competitiveness",
-      name: "Competitiveness",
-      description: "are competitive"
-    },
+    // TODO: Add competitiveness display (??)
+    // {
+    //   key: "competitiveness",
+    //   name: "Competitiveness",
+    //   description: "are competitive"
+    // },
     {
       key: "compactness",
       name: "Compactness",
       type: "percent",
       longText:
-        "A district that is more spread out has a lower compactness. Low compactness can be an indicators of gerrymandering when lorem ipsum lorem ipsum. Above XX% is considered good. Calculated using the Polsby-Popper method.",
+        "A district that efficiently groups constituents together has higher compactness. Low compactness or districts that branch out to different areas can be indicators of gerrymandering. Compactness is calculated using the Polsby-Popper method. Higher numbers are better.",
       shortText:
-        "A district that is more spread out has a lower compactness. Low compactness can be an indicators of gerrymandering",
+        "A district that is more spread out has a lower compactness. Low compactness can be an indicator of gerrymandering.",
       description: "are compact",
       value: avgCompactness
     },
-    {
-      key: "minorityMajority",
-      name: "Minority-Majority",
-      type: "fraction",
-      description: "are minority-majority",
-      shortText:
-        "A district in which the majority of the constituents are racial or ethnic minorities",
-      longText:
-        "A district in which the majority of the constituents are racial or ethnic minorities",
-      value: 1
-    },
+    // TODO: Minority-Majority updates (#750)
+    // {
+    //   key: "minorityMajority",
+    //   name: "Minority-Majority",
+    //   type: "fraction",
+    //   description: "are minority-majority",
+    //   shortText:
+    //     "A district in which the majority of the constituents are racial or ethnic minorities",
+    //   longText:
+    //     "A district in which the majority of the constituents are racial or ethnic minorities",
+    //   value: 1
+    // },
     {
       key: "countySplits",
       name: `${geoLevel ? geoLevelLabelSingular(geoLevel) : ""} splits`,
       type: "count",
       description: "are split",
-      shortText: "Lorem ipsum",
-      longText: "Lorem ipsum",
+      shortText:
+        "County splits occur when a county is split between two or more districts. Some states require minimizing county splits, to the extent practicable.",
+      longText:
+        "County splits occur when a county is split between two or more districts. Some states require minimizing county splits, to the extent practicable.",
       value: project?.districtsDefinition.filter(x => Array.isArray(x)).length || 0,
       total: project ? project.districtsDefinition.length : 0,
       splitCounties: project?.districtsDefinition.map(c => {